moved schemas to separate folder
