+++ utils/validation/addContactValidationSchema.js
@@ -1,0 +1,16 @@
+const Joi = require("joi");
+
+const validateAddContact = (req, res, next) => {
+  const schema = Joi.object({
+    name: Joi.string().min(2).required(),
+    email: Joi.string().email().required(),
+    phone: Joi.string().required(),
+  });
+  const validationResult = schema.validate(req.body);
+  if (validationResult.error) {
+    return res.status(400).json(validationResult.error.details);
+  }
+  next();
+};
+
+module.exports = { validateAddContact };
